use saveDocument from file server and simplify tool factory for agents
